DRY refactoring run
